Next refactoring, taking out more to cewIinfo, clipArt and a new imageUtils
